Update: RefreshToken Entity 제거

+++ back-end/src/main/java/com/exmaple/flory/util/SecurityUtil.java
@@ -10,13 +10,13 @@ public class SecurityUtil {
     private SecurityUtil() { }
 
     // Request 가 들어올 때 JwtFilter 의 doFilter 에서 저장
-    public static Long getCurrentMemberId() {
+    public static Integer getCurrentMemberId() {
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if (authentication == null || authentication.getName() == null) {
             throw  new RuntimeException("Security Context 에 인증 정보가 없습니다.");
         }
 
-        return Long.parseLong(authentication.getName());
+        return Integer.parseInt(authentication.getName());
     }
 }
